fix: Address review feedback on universal events pipeline
- Move stream.rs tests to separate stream_tests.rs file (CLAUDE.md convention)
- Change ItemStatus::InProgress to Completed for extracted tool calls
  (tool_use blocks are fully formed by drain time, InProgress misleads frontend)
- Remove duplicate agent_name from UniversalItem (carried at envelope level
  in UniversalItemsData, matching CoworkerStatusData pattern)
- Update stream.rs module doc to reflect broader scope
- Add universal events pipeline section to docs/architecture.md
